requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
numpy==2.3.3
    # via pandas
pandas==2.3.2
    # via torgstat (setup.py)
psycopg2-binary==2.9.10
    # via torgstat (setup.py)
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via torgstat (setup.py)
pytz==2025.2
    # via pandas
six==1.17.0
    # via python-dateutil
sqlalchemy==2.0.43
    # via torgstat (setup.py)
typing-extensions==4.15.0
    # via sqlalchemy
tzdata==2025.2
    # via pandas
